backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
# djongo==1.3.3  # Commented out for SQLite development
# pymongo==4.5.0

# Authentication & Security
django-allauth==0.57.0
python-decouple==3.8
cryptography==41.0.7

# File Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.29.7

# Communication
channels==4.0.0
channels-redis==4.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.4
redis==5.0.1

# Development
django-debug-toolbar==4.2.0
ipython==8.17.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# API Documentation
drf-spectacular==0.26.5

# Geolocation
geopy==2.4.0
